+++ src/app/(beta)/timetabling/units/page.tsx
@@ -1,0 +1,13 @@
+import { notFound } from "next/navigation";
+
+import { useFeatureFlag } from "@/utils/featureFlags";
+
+const Page = async () => {
+  const isEnabled = await useFeatureFlag("adopt-timetabling-proto", "boolean");
+  if (!isEnabled) {
+    return notFound();
+  }
+  return <p>View timetable</p>;
+};
+
+export default Page;