+++ CHANGE_LOG.md
@@ -1,3 +1,79 @@
+# [1.865.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.864.0...v1.865.0) (2025-09-23)
+
+
+### Features
+
+* bump oak-components to v1.147.0 ([d155737](https://github.com/oaknational/Oak-Web-Application/commit/d15573793007907e51b193f88f822687c5862665))
+
+# [1.864.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.863.0...v1.864.0) (2025-09-22)
+
+
+### Features
+
+* **timetabling:** scaffold UI with headings and disabled inputs ([858ff5b](https://github.com/oaknational/Oak-Web-Application/commit/858ff5b284b8f28aaac38e71a49defe033083933))
+
+# [1.863.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.862.1...v1.863.0) (2025-09-22)
+
+
+### Bug Fixes
+
+* remove hiding of campaign banner in browse journey ([c6dfdd6](https://github.com/oaknational/Oak-Web-Application/commit/c6dfdd635b5d47a6ce2870b21949aa504355559c))
+
+
+### Features
+
+* remove mythbusting feature flag ([7909f4d](https://github.com/oaknational/Oak-Web-Application/commit/7909f4d6285a5a634913de330da41e6602d739d8))
+
+## [1.862.1](https://github.com/oaknational/Oak-Web-Application/compare/v1.862.0...v1.862.1) (2025-09-22)
+
+
+### Bug Fixes
+
+* curric component storybooks ([04f78da](https://github.com/oaknational/Oak-Web-Application/commit/04f78dad254c9650e05b4c56f3ffc23b08612403))
+
+# [1.862.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.861.0...v1.862.0) (2025-09-22)
+
+
+### Features
+
+* added <CurricTimetableHeader/> component ([7dcfb8d](https://github.com/oaknational/Oak-Web-Application/commit/7dcfb8dcc20a3025958c3199624fc025b857e7d5))
+* added <CurricTimetableHeader/> into /timetabling/new page ([35b614d](https://github.com/oaknational/Oak-Web-Application/commit/35b614d90180bd195f21f079287fcf5ce198e117))
+
+# [1.861.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.860.0...v1.861.0) (2025-09-18)
+
+
+### Features
+
+* added <CurricShowSteps/> component ([c6bcd98](https://github.com/oaknational/Oak-Web-Application/commit/c6bcd98a6f42d14e69132ea79efb78f55f3be2bd))
+
+# [1.860.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.859.0...v1.860.0) (2025-09-18)
+
+
+### Features
+
+* make useFeatureFlag typesafe ([4dc7c7f](https://github.com/oaknational/Oak-Web-Application/commit/4dc7c7fcc23f70714b443f0c28bf6f542ee60529))
+* useFeatureFlag now just returns value and does type assertion on the type ([7d2efea](https://github.com/oaknational/Oak-Web-Application/commit/7d2efea0eba7e69a31bd78968b50a6e51d222b59))
+
+# [1.859.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.858.0...v1.859.0) (2025-09-18)
+
+
+### Bug Fixes
+
+* update tests to use renderWithTheme ([ddd729c](https://github.com/oaknational/Oak-Web-Application/commit/ddd729c0dcad7390f241f731c18738e659d0a358))
+
+
+### Features
+
+* add page view to page ([451eb09](https://github.com/oaknational/Oak-Web-Application/commit/451eb09311090ef2d7d0c020134641b3713dd9a5))
+* create page view ([4cb54be](https://github.com/oaknational/Oak-Web-Application/commit/4cb54beb9c7e7ab90d0791819b7d6e39a3110a5d))
+
+# [1.858.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.857.0...v1.858.0) (2025-09-17)
+
+
+### Features
+
+* add routing and pages ([3c339f0](https://github.com/oaknational/Oak-Web-Application/commit/3c339f0c197c033f2cf5cf42ac91fff6d342d621))
+
 # [1.857.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.856.1...v1.857.0) (2025-09-17)